Added validation to the login page

+++ client/src/login/TechnicianLogin.js
@@ -1,6 +1,14 @@
+//Importing React and it's modules
 import React from 'react';
-import './LoginPages.css';
+
+//Importing Axios in order to communicate with API
 import Axios from 'axios';
+
+//Importing Page Styling
+import './LoginPages.css';
+
+//Importing page validator
+import LoginPageValidator from './LoginPageValidator';
 
 export default class TechnicianLogin extends React.Component{
     constructor(props){
@@ -17,24 +25,31 @@ export default class TechnicianLogin extends React.Component{
         let pwd = document.getElementById('pwd-input').value;
         //Declaring variable used in checking whether the entered credentials match a existing technician account
         let matched = false;
-        //Making API call to check if the entered credentials exist in an account in the fdm database
-        Axios.post('http://localhost:3001/api/get/technician/credentials-check', {
-            id: id.toString(),
-            pwd: pwd.toString()
-          }).then((response) => {
-            /*A promise which checks the API response and if the credentials are a match with an existing accounts credentials 
-                creates the necessary cookie and calls parent components redirect function */
-            matched = response.data.match;
-            if (!matched) {document.getElementById('error-msg').innerHTML = "Sign in Failed: Incorrect Credentials"}
-            else {
-                document.getElementById('error-msg').innerHTML = "";
-                //Saving credentials in local storage
-                localStorage.setItem("id", id.toString());
-                localStorage.setItem("pwd", pwd.toString());
-                localStorage.removeItem("fdmEmail"); // removing fdm email key-value pair
-                this.state.loadPageContent();   
+        //Creating a validator object to validate the inputs
+        let validator = new LoginPageValidator(document.getElementById('error-msg'));
+
+        if (validator.fdmIdExists(id)){
+            if (validator.pwdExists(pwd)){
+                //Making API call to check if the entered credentials exist in an account in the fdm database
+                Axios.post('http://localhost:3001/api/get/technician/credentials-check', {
+                    id: id.toString(),
+                    pwd: pwd.toString()
+                }).then((response) => {
+                    /*A promise which checks the API response and if the credentials are a match with an existing accounts credentials 
+                        creates the necessary cookie and calls parent components redirect function */
+                    matched = response.data.match;
+                    if (!matched) {document.getElementById('error-msg').innerHTML = "Sign in Failed: Incorrect Credentials"}
+                    else {
+                        document.getElementById('error-msg').innerHTML = "";
+                        //Saving credentials in local storage
+                        localStorage.setItem("id", id.toString());
+                        localStorage.setItem("pwd", pwd.toString());
+                        localStorage.removeItem("fdmEmail"); // removing fdm email key-value pair
+                        this.state.loadPageContent();   
+                    }
+                });
             }
-          });
+        }
     }
 
     render(){
